feat(nix): add nixos-module and overlay
This commit introduces a nixos-module and overlay as a first step to
enabling deployments. A configurable PORT environment variable has been
introduced to both the api and client crates to help users avoid port
collisions when deploying to multi-tenant NixOS servers.

The next step will be to also make the SKIP_CHECKS and DRY_RUN flags
configurable by the user at the nixos-module level.